// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  student: 'student',
  coordinator: 'coordinator',
  admin: 'admin'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  active: 'active',
  inactive: 'inactive'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const ActivityCategory = {
  workshop: 'workshop',
  seminar: 'seminar',
  training: 'training',
  extracurricular: 'extracurricular'
} as const

export type ActivityCategory = (typeof ActivityCategory)[keyof typeof ActivityCategory]


export const ActivityStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed'
} as const

export type ActivityStatus = (typeof ActivityStatus)[keyof typeof ActivityStatus]


export const ApplicationStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected'
} as const

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus]


export const NotificationType = {
  approval: 'approval',
  rejection: 'rejection',
  announcement: 'announcement',
  reminder: 'reminder'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]


export const AttendanceStatus = {
  present: 'present',
  absent: 'absent'
} as const

export type AttendanceStatus = (typeof AttendanceStatus)[keyof typeof AttendanceStatus]